+++ sway-core/src/language/ty/declaration/struct.rs
@@ -12,6 +12,7 @@ use sway_types::{Ident, Named, Span, Spanned};
 use crate::{
     engine_threading::*,
     language::{CallPath, Visibility},
+    semantic_analysis::type_check_context::MonomorphizeHelper,
     transform,
     type_system::*,
 };
