+++ sway-core/src/language/ty/declaration/trait.rs
@@ -8,6 +8,7 @@ use crate::{
     },
     engine_threading::*,
     language::{parsed, Visibility},
+    semantic_analysis::type_check_context::MonomorphizeHelper,
     transform,
     type_system::*,
 };
