+++ sway-core/src/language/ty/declaration/trait_fn.rs
@@ -5,6 +5,7 @@ use sway_types::{Ident, Named, Span, Spanned};
 use crate::{
     engine_threading::*,
     language::{ty::*, Purity},
+    semantic_analysis::type_check_context::MonomorphizeHelper,
     transform,
     type_system::*,
 };
