+++ sway-core/src/semantic_analysis.rs
@@ -6,7 +6,7 @@ mod module;
 pub mod namespace;
 mod node_dependencies;
 mod program;
-mod type_check_context;
+pub(crate) mod type_check_context;
 pub use ast_node::*;
 pub use namespace::Namespace;
 pub(crate) use type_check_context::TypeCheckContext;
