+++ sway-core/src/semantic_analysis/ast_node/declaration/trait_fn.rs
@@ -3,6 +3,7 @@ use sway_types::{Span, Spanned};
 use crate::{
     decl_engine::DeclId,
     language::{parsed, ty, Visibility},
+    semantic_analysis::type_check_context::EnforceTypeArguments,
 };
 use sway_error::handler::{ErrorEmitted, Handler};
