+++ sway-core/src/type_system/priv_prelude.rs
@@ -6,7 +6,6 @@ pub(crate) use super::{
         create_type_id::CreateTypeId,
         replace_self_type::ReplaceSelfType,
     },
-    engine::{EnforceTypeArguments, MonomorphizeHelper},
     info::VecSet,
     substitute::{subst_list::SubstList, subst_map::TypeSubstMap, subst_types::SubstTypes},
     unify::unify_check::UnifyCheck,
